Add full text search usage

+++ NexAI.Zendesk/Queries/FindSimilarZendeskIssuesByNumberQuery.cs
@@ -4,7 +4,7 @@ namespace NexAI.Zendesk.Queries;
 
 public class FindSimilarZendeskIssuesByNumberQuery(Options options)
 {
-    public async Task<List<SimilarIssue>> Handle(string number, ulong limit)
+    public async Task<List<SimilarIssue>> Handle(string number, int limit)
     {
         var getZendeskIssueByNumberQuery = new GetZendeskIssueByNumberQuery(options);
         var zendeskIssue = await getZendeskIssueByNumberQuery.Handle(number);
@@ -14,7 +14,7 @@ public class FindSimilarZendeskIssuesByNumberQuery(Options options)
         var similarIssues = await findSimilarIssuesByPhraseQuery.Handle(zendeskIssue.CombinedContent(), limit + 1);
         return similarIssues
             .Where(issue => issue.Number != number)
-            .Take((int)limit)
+            .Take(limit)
             .ToList();
     }
 }
