+++ NexAI.Zendesk/Queries/FindSimilarZendeskIssuesByPhraseQuery.cs
@@ -10,11 +10,11 @@ public class FindSimilarZendeskIssuesByPhraseQuery(Options options)
     private readonly QdrantOptions _qdrantOptions = options.Get<QdrantOptions>();
     private readonly TextEmbedder _textEmbedder = new(options);
     
-    public async Task<List<SimilarIssue>> Handle(string phrase, ulong limit)
+    public async Task<List<SimilarIssue>> Handle(string phrase, int limit)
     {
         using var client = new QdrantClient(_qdrantOptions.Host, _qdrantOptions.Port);
         var embedding = await _textEmbedder.GenerateEmbedding(phrase);
-        var response = await client.SearchAsync(ZendeskIssueCollections.QdrantCollectionName, embedding, limit: limit);
+        var response = await client.SearchAsync(ZendeskIssueCollections.QdrantCollectionName, embedding, limit: (ulong)limit);
         return response
             .Select(point => new SimilarIssue(point.Payload["number"].StringValue, point.Score))
             .ToList();
